pyproject.toml
[project]
name = "10x"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.3.0",
    "ipykernel>=7.0.1",
    "matplotlib>=3.10.7",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "pip>=25.3",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "transformers>=4.57.1",
]
